+++ Ullmannite/src/UiElement/Controls/UiButton.cpp
@@ -4,9 +4,9 @@
 
 using namespace Ull;
 
-UiButton::UiButton(std::string name, glm::uvec2 position, glm::uvec2 size) :
+UiButton::UiButton(std::string name, glm::vec2 position, glm::vec2 size) :
     UiBasicControl(name, position, size, UiControlType::UiButton),
-    m_buttonText{ std::make_shared<UiText>(std::string(name + "Text"), glm::uvec2(0, 0), size, L"Button") }
+    m_buttonText{ std::make_shared<UiText>(std::string(name + "Text"), glm::vec2(0, 0), size, L"Button") }
 {
     m_buttonText->SetAlignment(UiText::HorizontalAlignment::CENTER, UiText::VerticalAlignment::CENTER);
     m_buttonText->SetSampleThreshold(1.5f);
@@ -15,7 +15,7 @@ UiButton::UiButton(std::string name, glm::uvec2 position, glm::uvec2 size) :
     AddChildNode(m_buttonText);
 }
 
-void UiButton::SetSize(const glm::uvec2 size)
+void UiButton::SetSize(const glm::vec2 size)
 {
     UiBasicControl::SetSize(size);
     m_buttonText->SetSize(size);
