+++ Ullmannite/src/UiElement/Controls/UiSpace.cpp
@@ -3,7 +3,7 @@
 
 namespace Ull
 {
-    UiSpace::UiSpace(const std::string name, const glm::uvec2 position, const glm::uvec2 size) :
+    UiSpace::UiSpace(const std::string name, const glm::vec2 position, const glm::vec2 size) :
         UiBasicControl(name, position, size, UiControlType::UiSpace)
     {
     }
